fix(personal-website): serve MCP at /mcp directly, not just /api/mcp
vercel.json rewrites "/" on the mcp.rainforest.tools host to /api/mcp, but
this doesn't reliably take effect: astro build's own generated
.vercel/output/config.json has no knowledge of the project's separate
vercel.json and already routes "/" to the SSR function itself, which
serves the normal homepage-locale-redirect for any unrecognized path
including "/". Empirically, requests to the bare mcp.rainforest.tools/
302 to /en/ instead of reaching the MCP handler.

Rather than fight Vercel's undocumented merge order between vercel.json
and framework-generated routes, adds a second, real Astro route at
/mcp that serves the identical MCP tool surface — reachable at
rainforest.tools/mcp on the main domain, through the exact same _render
function and routing mechanism /api/mcp already uses reliably. No
dependency on any host-based rewrite.

Extracts the MCP server-building logic (previously inline in
src/pages/api/mcp.ts) into createProfileMcpHandler(basePath) so both
routes share one implementation. mcp-handler validates each request's
pathname against a value it derives from basePath
(deriveEndpointsFromBasePath), so /api/mcp keeps basePath: '/api' and
the new /mcp route uses the library's own default ('/mcp'), matching
where each is actually mounted.
